requirements.txt
black==24.8.0
click==8.1.7
humanfriendly==10.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
PyYAML==6.0.1
tqdm==4.66.4
typing_extensions==4.12.2
